pyproject.toml
[project]
name = "cnru"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "FUJISHIGE TEMMA", email = "tenma.x0@gmail.com" }]
dependencies = [
    "jax>=0.4.13",
    "jaxlib>=0.4.13",
    "click>=8.1.7",
    "flax>=0.7.2",
    "optax>=0.1.7",
    "rerun-sdk>=0.9.1",
    "wandb>=0.15.12",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ruff>=0.0.290", "black>=23.9.1"]

[tool.ruff]
line-length = 150

[tool.hatch.metadata]
allow-direct-references = true
